:heavy_check_mark: feat (Section 2): Assignment 2 - Question 1
In the last Lecture Video, we Wrote `uapi_mm.h` file in which we defined
`MM_REG_STRUCT` macro. Through this assignment, we shall be going to test
whether Phase 1 of the project is bug-free and is functional as expected.

Write a function with below Signature, provide the function implementation
in `mm.c` file, and declaration in `uapi_mm.h` file.